fix: cast disk_total_space/disk_free_space to int to prevent TypeError 500 on /admin/backups
disk_total_space() and disk_free_space() return float per PHP spec for
large disks. With strict_types=1, passing float to formatBytes(int)
threw a TypeError -> HTTP 500 on the remote Termux/Linux server.

- BackupAdminController: explicit (int) cast on /
- BackupRepository: widen formatBytes() to int|float with internal cast

+++ app/Http/Controllers/Admin/BackupAdminController.php
@@ -31,13 +31,13 @@ class BackupAdminController extends Controller
 
         // Disk space info
         $backupPath = $this->backupRepo->getBackupDirectory();
-        $diskTotal = disk_total_space($backupPath) ?: 0;
-        $diskFree = disk_free_space($backupPath) ?: 0;
-        $diskUsed = $diskTotal - $diskFree;
+        $diskTotal = (int) (disk_total_space($backupPath) ?: 0);
+        $diskFree  = (int) (disk_free_space($backupPath)  ?: 0);
+        $diskUsed  = $diskTotal - $diskFree;
         $diskPercent = $diskTotal > 0 ? round(($diskUsed / $diskTotal) * 100, 1) : 0;
-        $formattedDiskUsed = $this->backupRepo->formatBytes((int) $diskUsed);
+        $formattedDiskUsed  = $this->backupRepo->formatBytes((int) $diskUsed);
         $formattedDiskTotal = $this->backupRepo->formatBytes((int) $diskTotal);
-        $formattedDiskFree = $this->backupRepo->formatBytes((int) $diskFree);
+        $formattedDiskFree  = $this->backupRepo->formatBytes((int) $diskFree);
 
         $scheduleInfo = [
             'daily_db'     => 'ทุกวัน เวลา 01:00 น.',
